refactor(server): extract persistence service
- isolate Supabase game/event persistence behind a dedicated
  service to reduce server.ts surface area without changing behavior
